Rename train_pt.py to train.py
