+++ setup.py
@@ -82,7 +82,7 @@ entry_points = {
         'sockeye-prepare-data = sockeye.prepare_data:main',
         'sockeye-quantize = sockeye.quantize:main',
         'sockeye-score = sockeye.score:main',
-        'sockeye-train = sockeye.train_pt:main',
+        'sockeye-train = sockeye.train:main',
         'sockeye-translate = sockeye.translate_pt:main',
         'sockeye-vocab = sockeye.vocab:main',
         'sockeye-rerank = sockeye.rerank:main',
